+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 6.23.4 (2024-09-24)
 * BraintreePayPal
   * Add PayPal vault edit flow (BETA)
-    * Add `BTPayPalClient.edit(_:completion:)` method that takes in a `BTPayPalVaultEditRequest` or a `BTPayPalVaultErrorHandlingEditRequest`
+    * Add `BTPayPalClient.edit(_:)` method that takes in a `BTPayPalVaultEditRequest` or a `BTPayPalVaultErrorHandlingEditRequest`
     * **Note:** This feature is currently in beta and may change or be removed in future releases
   * Send `isVaultRequest` for App Switch events to PayPal's analytics service (FPTI)
 * BraintreeShopperInsights (BETA)